Dodato:implementation

+++ SudBackend/src/main/java/rva/services/CrudService.java
@@ -7,13 +7,13 @@ import java.util.Optional;
 public interface CrudService <T>{
     List<T> getAll();
 
-    boolean existsById(long id);
+    boolean existsById(int id);
 
-    Optional<T> findById(long id);
+    Optional<T> findById(int id);
 
     T create(T t);
 
-    Optional<T> update(T t, long id);
+    Optional<T> update(T t, int id);
 
-    void delete(long id);
+    void delete(int id);
 }
